Initial Zookeeper configuration.
This is consuming metadata with a proposed version of Confd
on the backend. A couple of things need to be overcome, when
the service is started, the first two containers will not
have all of the server entries. You need to restart the service
in order to get a consistent config.